test fixes

+++ src/example-archive/SAW/working/00005.tutorial-double.c
@@ -23,18 +23,18 @@ print "Done.";
 */
 
 int double_ref(int x)
-  /*@ requires let prod = 2i64 * (i64)x;
-               -2147483648i64 <= prod; prod<2147483647i64; 
-      ensures return == (i32) prod; 
+  /*@ requires let prod = 2 * x;
+               -2147483648 <= prod; prod<2147483647; 
+      ensures return == prod; 
 @*/
 {
     return x * 2;
 }
 
 int double_imp(int x) 
-  /*@ requires let prod = 2i64 * (i64)x;
-               0i32 <= x; prod<2147483647i64; 
-      ensures return == (i32) prod; 
+  /*@ requires let prod = 2 * x;
+               0 <= x; prod<2147483647; 
+      ensures return == prod; 
       @*/
 {
   return x << 1;
