+++ src/example-archive/simple-examples/inprogress/power_3.c
@@ -2,38 +2,38 @@
 // TODO: fix this 
 
 /*@
-lemma LemmaPowerUFDef(i32 i)
+lemma LemmaPowerUFDef(integer i)
   requires 
-    0i32 <= i
+    0 <= i
   ensures 
-    power_uf(2i32,0i32) == 1i32; 
-    power_uf(2i32,i+1i32) == (power_uf(2i32,i) * 2i32)
+    power_uf(2,0) == 1; 
+    power_uf(2,i+1) == (power_uf(2,i) * 2)
 @*/
 
 /*@
-lemma LemmaPowerOrd(i32 i, i32 j)
+lemma LemmaPowerOrd(integer i, integer j)
   requires
-    0i32 <= i;
+    0 <= i;
     i < j
   ensures 
-    (power_uf(2i32, i) * 2i32) <= power_uf(2i32,j)
+    (power_uf(2, i) * 2) <= power_uf(2,j)
 @*/
 
 int power2_3(int y)
 /*@ requires 
-      let MAXi32 = 2147483647i64; 
-      0i32 < y; 
-      (i64) power_uf(2i32,y) <= MAXi32 @*/
-/*@ ensures return == power_uf(2i32,y) @*/
+      let MAXi32 = 2147483647; 
+      0 < y; 
+      power_uf(2,y) <= MAXi32 @*/
+/*@ ensures return == power_uf(2,y) @*/
 {
     int i = 0;
     int pow = 1;
     /*@ apply LemmaPowerUFDef(i); @*/
 
     while (i < y)
-    /*@ inv 0i32 <= i; i <= y;
+    /*@ inv 0 <= i; i <= y;
             {y}unchanged; 
-            pow == power_uf(2i32,i) @*/
+            pow == power_uf(2,i) @*/
     {
         /*@ apply LemmaPowerUFDef(i); @*/
         /*@ apply LemmaPowerOrd(i,y); @*/
