+++ src/example-archive/simple-examples/working/add_3.c
@@ -3,9 +3,9 @@
 
 signed int add_3(signed int x, signed int y)
 /*@ requires 
-      let MAXi32 = 2147483647i64; 
-      let MINi32 = -2147483648i64;
-      let sum = (i64) x + (i64) y;
+      let MAXi32 = 2147483647; 
+      let MINi32 = -2147483648;
+      let sum = x + y;
       MINi32 <= sum; sum <= MAXi32;
     ensures return == x + y; @*/
 {
