+++ src/example-archive/simple-examples/working/add_uint_1.c
@@ -3,8 +3,8 @@
 
 unsigned int add_uint_1(unsigned int x, unsigned int y)
 /*@ requires 
-      let MAXi32 = 2147483647i64;
-      (i64) x + (i64) y <= MAXi32;
+      let MAXi32 = 2147483647;
+      x + y <= MAXi32;
     ensures return == x + y; @*/
 {
   signed int i;
